requirements.txt
python-dotenv
fastapi
aiohttp
uvicorn